Sync png to svg
We just updated the svg's to remove a phantom mustache but some of the
png's still had it. I went through each updated svg and reexported it to
match the latest source svg.

Also one of the icon svg's was the svg for the badge, which has the blue
circle around Porter Cat's head. The original png didn't have that blue
circle, so I fixed the svg to match and exported that too.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>